+++ bactopia/databases/ncbi.py
@@ -57,6 +57,38 @@ def is_biosample(accession: str) -> bool:
     )
 
 
+def get_taxid_from_species(species: str) -> str:
+    """
+    Convert a species name into a tax_id
+
+    Args:
+        species (str): A species name
+
+    Returns:
+        str: The corresponding tax_id
+    """
+    r = requests.get(
+        f"https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=taxonomy&term={species}"
+    )
+    taxid = None
+    if r.status_code == requests.codes.ok:
+        for line in r.text.split("\n"):
+            line = line.strip()
+            if line.startswith("<Id>"):
+                taxid = line.replace("<Id>", "").replace("</Id>", "")
+        if taxid:
+            logging.debug(f"Found taxon ID ({taxid}) for {species}")
+            return taxid
+        else:
+            logging.error(
+                f"Unable to determine taxon ID from {species}, please check spelling or try again later."
+            )
+            sys.exit(1)
+    else:
+        logging.error("Unexpected error querying NCBI, please try again later.")
+        sys.exit(1)
+
+
 def taxid2name(taxids: list, ncbi_api_key: str, chunk_size: int) -> dict:
     """
     Convert a list of NCBI TaxIDs to species names.