+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.6.1
+
+- shuffle ncbi related module out of `utils` and into `ncbi`
+- fixed missing import in `bactopia-prepare`
+
+
 ## 1.6.0
 
 - `bactopia-search`